requirements.txt
asgiref==3.2.3
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
idna==2.9
lxml==4.5.0
numpy==1.18.2
Pillow==7.0.0
pylatexenc==2.1
python3-openid==3.1.0
pytz==2019.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
torch==1.4.0+cu92
torchvision==0.5.0+cu92
urllib3==1.25.8